requirements.txt
pandas
openai
psycopg
python-dotenv
timescale-vector
instructor
anthropic
langfuse
docling
docling-core
tiktoken
fastapi
uvicorn[standard]
python-multipart
celery[redis]
redis